backend/pyproject.toml
[project]
name = "intub-backend"
version = "0.1.0"
description = "Backend API for Intubation Difficulty Assessment"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
]
